✨ feat(mes): 修改流水号字段类型并更新相关逻辑
将流水号字段类型从 Integer 修改为 Long，以支持更大的序号范围。更新相关的增量生成逻辑，确保在 Redis 中正确处理初始值和步长。同时，调整测试用例以反映这一变化。

+++ yudao-module-mes/src/main/java/cn/iocoder/yudao/module/mes/service/md/autocode/strategy/MesMdAutoCodeContext.java
@@ -33,6 +33,6 @@ public class MesMdAutoCodeContext {
      *
      * 由流水号策略生成并设置，用于保存到生成记录表。例如：当规则包含流水号分段时，生成编码 "ITEM_20260304_0001"，此字段值为 1
      */
-    private Integer serialNo;
+    private Long serialNo;
 
 }
